+++ src/main/kotlin/io/kjson/JSON.kt
@@ -28,6 +28,7 @@ package io.kjson
 import java.math.BigDecimal
 
 import io.kjson.parser.Parser
+import net.pwall.json.JSONFunctions.displayString
 
 /**
  * JSON core library - a set of functions to assist with the creation and output of JSON data.
@@ -48,19 +49,9 @@ object JSON {
 
     fun parse(json: String): JSONValue? = Parser.parse(json)
 
-    fun parseObject(json: String): JSONObject {
-        val result = parse(json)
-        if (result !is JSONObject)
-            throw JSONException("JSON is not an object")
-        return result
-    }
+    fun parseArray(json: String): JSONArray = parse(json).asArray
 
-    fun parseArray(json: String): JSONArray {
-        val result = parse(json)
-        if (result !is JSONArray)
-            throw JSONException("JSON is not an array")
-        return result
-    }
+    fun parseObject(json: String): JSONObject = parse(json).asObject
 
     fun JSONValue?.toJSON(): String = this?.toJSON() ?: "null"
 
@@ -71,11 +62,71 @@ object JSON {
             appendTo(a)
     }
 
-    fun Appendable.appendJSON(json: JSONValue?) = apply {
+    fun Appendable.appendJSONValue(json: JSONValue?) = apply {
         if (json == null)
             append("null")
         else
             json.appendTo(this)
     }
+
+    fun JSONValue?.displayValue(): String {
+        return when (this) {
+            null -> "null"
+            is JSONString -> displayString(value, 21)
+            is JSONArray -> when (size) {
+                0 -> "[]"
+                1 -> "[${this[0].displayValue()}]"
+                else -> "[...]"
+            }
+            is JSONObject -> when (size) {
+                0 -> "{}"
+                1 -> entries.iterator().next().let { "{${displayString(it.key, 21)}:${it.value.displayValue()}}" }
+                else -> "{...}"
+            }
+            else -> toString()
+        }
+    }
+
+    val JSONValue?.asString: String
+        get() = asStringOrNull ?: throw JSONException("Not a string - ${displayValue()}")
+
+    val JSONValue?.asStringOrNull: String?
+        get() = (this as? JSONString)?.value
+
+    val JSONValue?.asInt: Int
+        get() = asIntOrNull ?: throw JSONException("Not an int - ${displayValue()}")
+
+    val JSONValue?.asIntOrNull: Int?
+        get() = (this as? JSONNumberValue)?.let { if (it.isInt()) it.toInt() else null }
+
+    val JSONValue?.asLong: Long
+        get() = asLongOrNull ?: throw JSONException("Not a long - ${displayValue()}")
+
+    val JSONValue?.asLongOrNull: Long?
+        get() = (this as? JSONNumberValue)?.let { if (it.isLong()) it.toLong() else null }
+
+    val JSONValue?.asDecimal: BigDecimal
+        get() = asDecimalOrNull ?: throw JSONException("Not a decimal - ${displayValue()}")
+
+    val JSONValue?.asDecimalOrNull: BigDecimal?
+        get() = (this as? JSONNumberValue)?.toDecimal()
+
+    val JSONValue?.asBoolean: Boolean
+        get() = asBooleanOrNull ?: throw JSONException("Not a boolean - ${displayValue()}")
+
+    val JSONValue?.asBooleanOrNull: Boolean?
+        get() = (this as? JSONBoolean)?.value
+
+    val JSONValue?.asArray: JSONArray
+        get() = asArrayOrNull ?: throw JSONException("Not an array - ${displayValue()}")
+
+    val JSONValue?.asArrayOrNull: JSONArray?
+        get() = (this as? JSONArray)
+
+    val JSONValue?.asObject: JSONObject
+        get() = asObjectOrNull ?: throw JSONException("Not an object - ${displayValue()}")
+
+    val JSONValue?.asObjectOrNull: JSONObject?
+        get() = (this as? JSONObject)
 
 }