+++ src/test/kotlin/io/kjson/JSONValueTest.kt
@@ -26,57 +26,13 @@
 package io.kjson
 
 import kotlin.test.Test
-import kotlin.test.assertFailsWith
-import kotlin.test.assertTrue
 import kotlin.test.expect
 
-import java.math.BigDecimal
-
-import io.kjson.JSON.appendJSON
+import io.kjson.JSON.appendJSONValue
 import io.kjson.JSON.appendTo
 import io.kjson.JSON.toJSON
 
 class JSONValueTest {
-
-    @Test fun `should create values using JSON object`() {
-        val testInt = JSON.of(54321)
-        expect(JSONInt(54321)) { testInt }
-        val testLong = JSON.of(2233445566778899)
-        expect(JSONLong(2233445566778899)) { testLong }
-        val testDecimal = JSON.of(BigDecimal("99.999"))
-        expect(JSONDecimal(BigDecimal("99.999"))) { testDecimal }
-        val testString = JSON.of("Hello!")
-        expect(JSONString("Hello!")) { testString }
-    }
-
-    @Test fun `should parse using JSON object`() {
-        val json = JSON.parse("""{"one":1,"two":2}""")
-        assertTrue(json is JSONObject)
-        expect(2) { json.size }
-        expect(JSONInt(1)) { json["one"] }
-        expect(JSONInt(2)) { json["two"] }
-    }
-
-    @Test fun `should parse using JSON object parseObject`() {
-        val json = JSON.parseObject("""{"one":1,"two":2}""")
-        expect(2) { json.size }
-        expect(JSONInt(1)) { json["one"] }
-        expect(JSONInt(2)) { json["two"] }
-        assertFailsWith<JSONException> { JSON.parseObject("[1,2,3]") }.let {
-            expect("JSON is not an object") { it.message }
-        }
-    }
-
-    @Test fun `should parse using JSON object parseArray`() {
-        val json = JSON.parseArray("""["alpha","beta","gamma"]""")
-        expect(3) { json.size }
-        expect(JSONString("alpha")) { json[0] }
-        expect(JSONString("beta")) { json[1] }
-        expect(JSONString("gamma")) { json[2] }
-        assertFailsWith<JSONException> { JSON.parseObject("[1,2,3]") }.let {
-            expect("JSON is not an object") { it.message }
-        }
-    }
 
     @Test fun `should use nullable functions`() {
         var testNull: JSONValue? = null
@@ -92,7 +48,7 @@ class JSONValueTest {
 
     @Test fun `should use Appendable appendJSON`() {
         val test = StringBuilder()
-        test.appendJSON(JSONArray.of(JSONInt(123), JSONInt(321)))
+        test.appendJSONValue(JSONArray.of(JSONInt(123), JSONInt(321)))
         expect("[123,321]") { test.toString() }
     }